backend/requirements.txt
fastapi>=0.93.0
uvicorn>=0.15.0
python-multipart>=0.0.5
openai>=0.27.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
sqlalchemy>=2.0.23
alembic>=1.13.0
psycopg2-binary>=2.9.9
pydantic>=1.10.0,<2.0.0
python-dotenv>=0.19.0
google-cloud-speech>=2.11.1
google-cloud-translate>=3.7.4
prefect>=2.10.0
pytest>=7.3.1
httpx>=0.24.1
